// src/app/models/card-name.enum.ts
export enum CardName {
  bubble = 'bubble dialog',
  draw = 'draw dialog',
  emotion = 'emotion dialog',
  memory = 'memory dialog',
  swim1 = 'swim1 dialog',
  swim2 = 'swim2 dialog',
  swim3 = 'swim3 dialog',
  swim4 = 'swim4 dialog',
  swim = 'swim dialog',
  toys = 'toys dialog',
  skiing = 'skiing dialog',
  athletics = 'athletics dialog',
  football = 'football dialog',
  gymnastics = 'gymnastics dialog',
  volleyball = 'volleyball dialog',
  tip1 = 'tip1 dialog',
  tip2 = 'tip2 dialog',
  tip3 = 'tip3 dialog',
  tip4 = 'tip4 dialog',
  tip5 = 'tip5 dialog',
  tip6 = 'tip6 dialog',
  tip7 = 'tip7 dialog',
  tip8 = 'tip8 dialog'
}